feat(FTO Backup): Add backup worker Details onBlur of Input Element + User Already Exists Validation
Currently, only on pressing Enter, the value is getting added as Chip element. With this change, user should be able to add Worker details by simplying removing the focus in TEXT Element. Moreover, a validation has been added to errorout if worker is already added. Third change is to ensure Backup Worker ACTIVE indicator is turn off if no worker is selected.

#Closes CCTP-13450 issue#1 and issue#3